requirements.txt
argparse
collections
configparser
csv
datetime
json
lxml
math
matplotlib
os
pandas
psycopg2
re
requests
seaborn
sklearn.linear_model
sklearn.metrics
sqlalchemy
sqlalchemy-utils
statistics
subprocess
sys
tenacity
time
tldextract
urllib.parse
xml.etree.ElementTree
